Centralize load_env in utils
