+++ logfile_etl.py
@@ -7,19 +7,7 @@ from urllib.parse import urlparse, parse_qs
 
 import paramiko
 from db_utils import AccessLogDB
-
-def load_env(path=".env"):
-    """Load key=value pairs from a .env file into os.environ"""
-    if not os.path.exists(path):
-        return
-    with open(path) as f:
-        for line in f:
-            line = line.strip()
-            if not line or line.startswith('#'):
-                continue
-            if '=' in line:
-                key, value = line.split('=', 1)
-                os.environ.setdefault(key, value)
+from utils import load_env
 
 load_env()
