+++ utils.py
@@ -1,5 +1,20 @@
+import os
 from datetime import datetime, timedelta, date
 from flask import request, render_template, url_for
+
+
+def load_env(path: str = ".env") -> None:
+    """Load key=value pairs from a .env file into ``os.environ``."""
+    if not os.path.exists(path):
+        return
+    with open(path) as f:
+        for line in f:
+            line = line.strip()
+            if not line or line.startswith("#"):
+                continue
+            if "=" in line:
+                key, value = line.split("=", 1)
+                os.environ.setdefault(key, value)
 
 def render_dashboard(template, tab, params, filter_from, filter_to, **kwargs):
     """
